Remove unused session sidebar header color registrations from chatColors.ts
