+++ src/vs/sessions/common/theme.ts
@@ -1,0 +1,34 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { localize } from '../../nls.js';
+import { registerColor } from '../../platform/theme/common/colorUtils.js';
+import { SIDE_BAR_BACKGROUND, SIDE_BAR_FOREGROUND } from '../../workbench/common/theme.js';
+
+// Sessions sidebar header colors
+export const sessionsSidebarHeaderBackground = registerColor(
+	'sessionsSidebarHeader.background',
+	SIDE_BAR_BACKGROUND,
+	localize('sessionsSidebarHeader.background', 'Background color of the sidebar header area in the agent sessions window.')
+);
+
+export const sessionsSidebarHeaderForeground = registerColor(
+	'sessionsSidebarHeader.foreground',
+	SIDE_BAR_FOREGROUND,
+	localize('sessionsSidebarHeader.foreground', 'Foreground color of the sidebar header area in the agent sessions window.')
+);
+
+// Chat bar title colors
+export const chatBarTitleBackground = registerColor(
+	'chatBarTitle.background',
+	SIDE_BAR_BACKGROUND,
+	localize('chatBarTitle.background', 'Background color of the chat bar title area in the agent sessions window.')
+);
+
+export const chatBarTitleForeground = registerColor(
+	'chatBarTitle.foreground',
+	SIDE_BAR_FOREGROUND,
+	localize('chatBarTitle.foreground', 'Foreground color of the chat bar title area in the agent sessions window.')
+);
